requirements.txt
Django==1.4.3
Jinja2==2.6
Pygments==1.5
South==0.7.6
Sphinx==1.1.3
amqp==1.0.6
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.19
boto==2.6.0
celery==3.0.12
distribute==0.6.32
dj-database-url==0.2.1
django-celery==3.0.11
django-crispy-forms==1.2.0
django-debug-toolbar==0.9.4
django-pylibmc-sasl==0.2.4
django-storages==1.1.5
docutils==0.9.1
gevent==0.13.8
greenlet==0.4.0
gunicorn==0.16.1
kombu==2.5.3
mock==1.0.1
newrelic==1.8.0.13
psycopg2==2.4.5
py-bcrypt==0.2
pylibmc==1.2.3
python-dateutil==1.5
pytz==2012h
rauth==0.4.17
requests==0.14.2
wsgiref==0.1.2
